backend/requirements.txt
Flask>=2.3.0
Pillow>=10.0.0
PyYAML>=6.0.1
accelerate>=0.25.0
av>=11.0.0
datasets>=2.15.0
dtw-python>=1.3.0
evaluate>=0.4.0
flask-cors>=4.0.0
flask-socketio>=5.3.6
huggingface_hub>=0.17.0
imageio-ffmpeg>=0.4.9
imageio>=2.31.0
jsonlines>=4.0.0
matplotlib>=3.7.0
numpy>=1.24.0,<2.0.0
opencv-python>=4.8.0
pandas>=2.0.0
plotly>=5.17.0
pyarrow>=13.0.0
pydantic>=2.5.0
pyserial>=3.5
pytest>=7.4.0
safetensors>=0.4.0
scikit-learn>=1.3.0
scipy>=1.11.0
seaborn>=0.12.0
sentence-transformers>=3.0.0,<3.1.0
spacy>=3.7.0
tokenizers>=0.15.0
torch>=2.4.0,<2.5.0
tqdm>=4.66.0
transformers>=4.53.0,<4.54.0
ultralytics>=8.0.0

# Authentication dependencies
PyJWT>=2.8.0
bcrypt>=4.1.2
# SQLAlchemy removed - using Firebase Firestore instead
python-dotenv>=1.0.0

# Firebase dependencies
firebase-admin>=6.4.0
